Dungeon fix

+++ BetterVanillaWorldGen/Interface/VanillaInterface.Static.cs
@@ -7,14 +7,14 @@ namespace AdvancedWorldGen.BetterVanillaWorldGen.Interface
 		
 		public VanillaAccessor<bool> SkipFramingDuringGen;
 		
-		public VanillaAccessor<uint> CrackedType;
+		public VanillaAccessor<ushort> CrackedType;
 
 		public void InitializeStatics()
 		{
 			NumOceanCaveTreasure = new VanillaAccessor<int>("numOceanCaveTreasure");
 			HeartCount = new VanillaAccessor<int>("heartCount");
 			SkipFramingDuringGen = new VanillaAccessor<bool>("skipFramingDuringGen");
-			CrackedType = new VanillaAccessor<uint>("crackedType");
+			CrackedType = new VanillaAccessor<ushort>("crackedType");
 		}
 	}
 }
